Moved testing to separate file, added histograms
